+++ test/unit/specs/store/session.spec.js
@@ -20,8 +20,6 @@ describe(`Module: Session`, () => {
       Sentry: {
         init: jest.fn()
       },
-      enableGoogleAnalytics: jest.fn(),
-      disableGoogleAnalytics: jest.fn(),
       track: jest.fn(),
       config: {
         development: false,
@@ -271,7 +269,6 @@ describe(`Module: Session`, () => {
 
     expect(state.errorCollection).toBe(true)
     expect(localStorage.getItem(`voyager_error_collection_abc`)).toBe(`true`)
-    expect(state.externals.enableGoogleAnalytics).toHaveBeenCalledWith(`UA-123`)
     expect(state.externals.track).toHaveBeenCalledWith(`pageview`, {
       dl: `/`
     })
@@ -294,9 +291,6 @@ describe(`Module: Session`, () => {
 
     expect(state.errorCollection).toBe(false)
     expect(localStorage.getItem(`voyager_error_collection_abc`)).toBe(`false`)
-    expect(state.externals.disableGoogleAnalytics).toHaveBeenCalledWith(
-      `UA-123`
-    )
     expect(state.externals.Sentry.init).toHaveBeenCalledWith({})
   })
 
@@ -318,9 +312,6 @@ describe(`Module: Session`, () => {
     })
     expect(state.errorCollection).toBe(false)
     expect(localStorage.getItem(`voyager_error_collection_abc`)).toBe(`false`)
-    expect(state.externals.disableGoogleAnalytics).toHaveBeenCalledWith(
-      `UA-123`
-    )
     expect(state.externals.Sentry.init).toHaveBeenCalledWith({})
   })